CSS fixed so the bug icon shows now
Apparently the @CHARSET is not being removed by the DokuWiki CSS
minimizer and both Chrome and Safari can't understand the resulting
CSS. Explains why it only worked in Firefox. I'll file a bug report for
DokuWiki.

And the path, which is corrected by the same CSS minimizer, is now
correct. I was overthinking it. ;-/